+++ LuckyoneUI/Core/Defaults.lua
@@ -160,6 +160,7 @@ Private.Defaults = {
 			autoAcceptRole = false,
 			autoDismount = false,
 			easyDelete = false,
+			expandMerchant = false,
 			preventLootAutoShow = false,
 			privacyOverlay = false,
 			quickSignup = false,