+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ﻿## Version 4.23 [In development, not yet released]
 
 **All WoW Versions:**  
+- Added option to double the amount of icons on vendor frames (4 rows of 5 items per page)  
 - Alt setup button is back, one click to load all existing profiles and optionally setup chat  
 
 **Retail:**  
 - Updated aura indicators for all healing specs (Size, Postion, Stack count, Style)  
+- Updated WindTools profile to disable extended merchant page (I use my own option for it now)  
 
 **Damage Meter:**  
 - Bookmarks on very small windows are now fixed and let you scroll if needed